renaming: _3D => ThreeD
